requirements.txt
Flask
flask-cors
requests
beautifulsoup4
selenium
webdriver-manager
google-generativeai
gunicorn